requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# gui.py: 5
Pillow == 8.4.0

# gui.py: 4
pywhatkit == 5.4
